Move pytest subdir into main test dir
