--- test/pytest/__init__.py
@@ -1,1 +1,0 @@
-# Nothing special about this directory, it's just to separate old/new tests
